refactor: assets routes
