+++ .vscode/settings.json
@@ -5,6 +5,7 @@
     "cardano",
     "cbor",
     "delegators",
+    "emurgo",
     "fastify",
     "milkomeda",
     "mirs",
